requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djangorestframework_simplejwt==4.5.0
pytest==6.2.4
